// src/app/game/enums/game-phase.enum.ts
/**
 * Создается лобби
 *
 * Фаза `prepare`
 *
 * Сервер ожидает запуска игры
 *
 * Игра запускается с кнопки
 * Фаза меняется
 *
 * Фаза `choose-situation`
 *
 * Запустился таймер для перехода к следующей фазе (обычный)
 * Сервер слушает события отправки голосов за ситуацию
 *
 * Фаза `situation`
 *
 * Запустился таймер для перехода к следующей фазе (выбранный при создании лобби)
 * Сервер слушает события отправки мемов
 *
 * При получении мемов происходит проверка "каждый ли игрок отправил мем"
 * При получении мемов от каждого игрока фаза меняется.
 * Если таймер закончился раньше получения всех мемов - фаза меняется
 *
 * Фаза `vote`
 *
 * Запустился таймер для перехода к следующей фазе (выбранный при создании лобби)
 * Сервер слушает события отправки голосов за мем
 *
 * При получении голосов происходит проверка "каждый ли игрок проголосовал"
 * При получении голосов от каждого игрока фаза меняется.
 * Если таймер закончился раньше получения всех мемов - фаза меняется
 *
 * Фаза `vote-results`
 *
 * Запустился таймер для перехода к следующей фазе (короткий)
 *
 * Если таймер закончился, то происходит проверка "последний ли раунд"
 * Если раунд не последний, переходим к фазе `situation` и начинается новый раунд
 * Если раунд последний, переходим к фазе `end`
 *
 * Фаза `end`
 *
 * TBD
 */

export enum GamePhase {
  PREPARE = 'prepare',
  CHOOSE_SITUATION = 'choose-situation',
  SITUATION = 'situation',
  VOTE = 'vote',
  VOTE_RESULTS = 'vote-results',
  END = 'end',
}